Move generic helper functions from api package to helper package.
