+++ helper/funcs.go
@@ -1,4 +1,4 @@
-package api
+package helper
 
 import (
 	"regexp"
